+++ CHANGELOG.md
@@ -5,18 +5,30 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
 
 ### Changed
-- Uses OpenSSL 1.0.2u to support FIPS compliance
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+## [1.8.0] - 2020-07-10
+### Changed
+- Use OpenSSL 1.0.2u to support FIPS compliance
   ([cyberark/conjur#1527](https://github.com/cyberark/conjur/issues/1527))
 - Conjur can be configured to run in FIPS compliant or Non-FIPS compliant mode depending on requirements.
   FIPS Compliant mode is slightly slower then non-FIPS compliant
   ([cyberark/conjur#1527](https://github.com/cyberark/conjur/issues/1527))
-- Bump conjur-rack from 4.0.0 to 4.2.0 that consumes FIPS compliant slosilo(https://github.com/cyberark/conjur/issues/1527))
-- Print login and authentication error to the log in INFO level.
+- Bump conjur-rack from 4.0.0 to 4.2.0 that consumes FIPS compliant slosilo([cyberark/conjur#1527](https://github.com/cyberark/conjur/issues/1527))
+- Print login and authentication error to the log in INFO level
   [cyberark/conjur#1377](https://github.com/cyberark/conjur/issues/1377)
 - Print proper message when user does not exist in authn or login request with
-  default authenticator.
+  default authenticator
   [cyberark/conjur#1655](https://github.com/cyberark/conjur/issues/1655)
 
 ### Added
@@ -27,7 +39,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - All audit events now contain the IP address of the client that initiated the
   API request (e.g. `[client@43868 ip="172.24.0.5"]`)
   ([cyberark/conjur#1550](https://github.com/cyberark/conjur/issues/1550))
-- Print Conjur server FIPS mode status. [cyberark/conjur#1654](https://github.com/cyberark/conjur/issues/1654)
+- Print Conjur server FIPS mode status ([cyberark/conjur#1654](https://github.com/cyberark/conjur/issues/1654))
+
+### Security
+- Updated `rack` to `2.2.3` to resolve CVE-2020-8184
 
 ## [1.7.4] - 2020-06-17
 
@@ -376,7 +391,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - The first tagged version.
 
-[Unreleased]: https://github.com/cyberark/conjur/compare/v1.7.4...HEAD
+[Unreleased]: https://github.com/cyberark/conjur/compare/v1.8.0...HEAD
+[1.8.0]: https://github.com/cyberark/conjur/compare/v1.7.4...v1.8.0
 [1.7.4]: https://github.com/cyberark/conjur/compare/v1.7.3...v1.7.4
 [1.7.3]: https://github.com/cyberark/conjur/compare/v1.7.2...v1.7.3
 [1.7.2]: https://github.com/cyberark/conjur/compare/v1.7.1...v1.7.2